[chore] : rename ui package to view package
