+++ src/main/java/view/InputView.java
@@ -1,4 +1,4 @@
-package ui;
+package view;
 
 import java.util.Scanner;
