+++ src/main/java/view/ResultView.java
@@ -1,4 +1,4 @@
-package ui;
+package view;
 
 import domain.RacingCar;
